(chore) move subcomponents to components/Review/sub, keep only controller ReviewPage at top-level
